# Coalesce timers to improve relay performance