This diff replaces the relayer's use of the standard lib's `time.Timer` with a custom timer wheel. The wheel reduces the overhead of maintaining timers in four ways:

- First, it stores timers in multiple buckets, which reduces lock contention.
- Within buckets, it uses a linked list instead of a heap, which allows for O(1) insertion.
- Second, it executes callback functions in the ticker thread (instead of spawning a goroutine per ticker).
- Lastly, it reduces timer cancellation to an atomic CAS.

The tradeoffs are, of course, that we coarsen timers to fire only every ~5ms, and we run a greater risk of blocking the tick thread when callbacks block.

With 10k relayed calls in flight, benchmarks suggest that this implementation is 50% faster than the standard lib:

```
# Schedule and immediately cancel a single timer.
BenchmarkScheduleAndCancelWheelNoHeap-4                 10000000               144 ns/op              48 B/op          1 allocs/op
BenchmarkScheduleAndCancelStandardLibraryNoHeap-4       10000000               239 ns/op              64 B/op          1 allocs/op

# Schedule 10k timers for tomorrow, then benchmark scheduling and canceling an additional timer.
BenchmarkScheduleAndCancelWheelWithHeap-4               10000000               156 ns/op              48 B/op          1 allocs/op
BenchmarkScheduleAndCancelStandardLibraryWithHeap-4      5000000               317 ns/op         
```

/cc @prashantv @billf @monicato @witriew