@akshayjshah 2016-12-03
Yep, I'll add one. I'll need to pivot some of the APIs to accommodate
benchmarking that part well.
On Sat, Dec 3, 2016 at 10:34 AM Prashant Varanasi <notifications@github.com>
wrote:

> *@prashantv* commented on this pull request.
>
> In timers/timers.go <https://github.com/uber/tchannel-go/pull/545>:
>
> > +	state    atomic.Int32
> +	prev     *Timer
> +}
> +
> +// Stop FIXME
> +func (t *Timer) Stop() {
> +	t.state.Store(stateCanceled)
> +}
> +
> +func (t *Timer) append(tail *Timer) *Timer {
> +	tail.prev = t
> +	return tail
> +}
> +
> +// Wheel FIXME
> +type Wheel struct {
>
> My bad on number of buckets, it's definitely a lot more memory to do that.
> We'll need to use a linked-list like approach, although each list element
> could store multiple timers.
>
> We need another benchmark that actually involves firing timers, since the
> optimizations I'm suggesting are optimizing the fired timers path, and the
> current benchmark only looks at creating+stopping timers.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/uber/tchannel-go/pull/545>, or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AA7X9ioAAN5ecLQLmFzNNtZCtNBvk2pCks5rEbZCgaJpZM4LCSfx>
> .
>
-- 
Akshay Shah
Software Engineer, RPC
shah@uber.com