@akshayjshah 2016-12-06
Updated with a singly-linked list implementation, which is ~twice as fast as the standard lib when scheduling and canceling a timer, and modestly faster when firing timeouts (it's much harder to accurately benchmark the latter case):

```
BenchmarkTimerCanceled-4                        20000000               104 ns/op              32 B/op          1 allocs/op
BenchmarkStandardLibraryTimerCanceled-4         10000000               221 ns/op              64 B/op          1 allocs/op

# There's 1ms of wait time in this benchmark, so we're only spending ~106ns firing our timers.
BenchmarkTimerWorkThread-4                          2000           1000106 ns/op              50 B/op          2 allocs/op

# Most of the time and allocs in these benchmarks are from the fake clock.
BenchmarkTimerExpiry-4                               100          14240678 ns/op            1104 B/op         24 allocs/op
BenchmarkStandardLibraryTimerExpiry-4                100          14342143 ns/op            1132 B/op         23 allocs/op
```