@akshayjshah 2016-12-08
Updated, now with fewer bit manipulation errors.

```
# Schedule and immediately cancel a single timer.
BenchmarkTimerScheduleAndCancelNoHeap-4                 10000000               144 ns/op              48 B/op          1 allocs/op
BenchmarkTimerScheduleAndCancelWithHeap-4               10000000               156 ns/op              48 B/op          1 allocs/op

# Schedule 10k timers for tomorrow, then benchmark scheduling and canceling an additional timer.
BenchmarkScheduleAndCancelStandardLibraryNoHeap-4       10000000               239 ns/op              64 B/op          1 allocs/op
BenchmarkScheduleAndCancelStandardLibraryWithHeap-4      5000000               317 ns/op              64 B/op          1 allocs/op
```