@akshayjshah 2016-12-15
Ready for another round of review. Planning follow-on PRs for:

- exporting an atomic to track whether the ticker thread is lagging;
- taking an interface instead of a function closure in `AfterFunc`;
- (possibly) putting backpressure on callers when the ticker thread lags.