@akshayjshah 2016-12-15
Tests are failing on Go tip, and I'm not quite sure why: https://travis-ci.org/uber/tchannel-go/jobs/184112308#L939

These tests and the code they exercise looks correct to me, and passes with `-count 10000` on my local Go install. Will look in more detail later.