@akshayjshah 2016-12-15
Fun story: I think we're putting too much load on our Travis containers. Tests are failing with no error message beyond "signal: killed", which sounds like https://docs.travis-ci.com/user/common-build-problems/#My-build-script-is-killed-without-any-error.