@akshayjshah 2016-12-19
Git surgery complete - unrelated test fixes are in #551. Once that's merged and this PR is rebased, tests should pass.