@akshayjshah 2017-06-20
This degree of perf doesn't seem to be necessary, so there's virtually no chance we'll spend the time required to get to the bottom of these Travis failures. Closing.